refactor(mobile): extract mouse report cell mapping
